attempt to make it buildable on cloud
