+++ diffcalculia.ts
@@ -1,7 +1,3 @@
-#!/usr/bin/env -S node --import tsx/esm
-
-import { stdin, stdout, stderr } from 'process';
-
 interface HunkFix {
   index: number;
   oldStart: number;
@@ -11,7 +7,6 @@ interface HunkFix {
 }
 
 export function validatePatch(patch: string, fixMode = false): string {
-  // Normalize newlines and drop a trailing empty line (like Python splitlines())
   const raw = patch.replace(/\r\n/g, '\n').split('\n');
   const lines = raw[raw.length - 1] === '' ? raw.slice(0, -1) : raw;
 
@@ -60,7 +55,6 @@ export function validatePatch(patch: string, fixMode = false): string {
   }
 
   if (fixMode && fixes.length) {
-    stderr.write('Let me fix that for you\n');
     for (const f of fixes) {
       lines[f.index] =
         `@@ -${f.oldStart},${f.actualOld} +${f.newStart},${f.actualNew} @@`;
@@ -69,40 +63,4 @@ export function validatePatch(patch: string, fixMode = false): string {
   }
 
   return patch;
-}
-
-function parseArgs() {
-  const args = process.argv.slice(2);
-  const idx = args.indexOf('--fix');
-  const fixMode = idx !== -1;
-  if (fixMode) args.splice(idx, 1);
-  return { fixMode };
-}
-
-function readStdin(): Promise<string> {
-  return new Promise((resolve, reject) => {
-    let data = '';
-    stdin.setEncoding('utf8');
-    stdin.on('data', chunk => data += chunk);
-    stdin.on('end', () => resolve(data));
-    stdin.on('error', reject);
-  });
-}
-
-async function main() {
-  const { fixMode } = parseArgs();
-  const patchText = await readStdin();
-  try {
-    const out = validatePatch(patchText, fixMode);
-    stdout.write(out);
-    stderr.write('✅ Patch validation passed\n');
-    process.exit(0);
-  } catch (err: any) {
-    stderr.write(`❌ ${err.message}\n`);
-    process.exit(1);
-  }
-}
-
-if (require.main === module) {
-  main();
 }
