+++ package.json
@@ -1,17 +1,22 @@
 {
   "name": "diffcalculia-ts",
   "version": "1.0.0",
-  "main": "dist/diffcalculia.js",
+  "main": "diffcalculia.ts",
+  "types": "diffcalculia.ts",
+  "bin": {
+    "diffcalculia": "./cli.ts"
+  },
   "scripts": {
     "test": "jest",
     "build": "tsc",
-    "prepublish": "npm run build"
+    "prepublishOnly": "npm run build"
   },
   "devDependencies": {
     "@types/jest": "^29.5.12",
     "jest": "^29.7.0",
     "ts-jest": "^29.1.1",
-    "typescript": "^5.3.3"
+    "typescript": "^5.3.3",
+    "tsx": "^4.19.4"
   },
   "jest": {
     "preset": "ts-jest",
@@ -20,7 +25,5 @@
       "**/tests/**/*.test.ts"
     ]
   },
-  "dependencies": {
-    "tsx": "^4.19.4"
-  }
+  "dependencies": {}
 }
